Fixing implicit cast warnings (2.rc.04)
Merge pull request by https://github.com/Celemony and
https://github.com/mynameisjohn

+++ base64.h
@@ -1,6 +1,6 @@
 //
 //  base64 encoding and decoding with C++.
-//  Version: 2.rc.03 (release candidate)
+//  Version: 2.rc.04 (release candidate)
 //
 
 #ifndef BASE64_H_C0CE2A47_D10E_42C9_A27C_C883944E704A
@@ -17,7 +17,7 @@ std::string base64_encode_pem (std::string const& s);
 std::string base64_encode_mime(std::string const& s);
 
 std::string base64_decode(std::string const& s, bool remove_linebreaks = false);
-std::string base64_encode(unsigned char const*, unsigned int len, bool url = false);
+std::string base64_encode(unsigned char const*, size_t len, bool url = false);
 
 #if __cplusplus >= 201703L
 //
